Implement early stopping (#56)
* Add early stopping

* Use early stopping baseline

* Add early stopping baseline results

* Fix early stopping on ddp

* Fix distributed

---------

+++ src/train.py
@@ -6,6 +6,7 @@ import pathlib
 import torch
 import tqdm
 from peft import set_peft_model_state_dict
+from torch.distributed import ReduceOp
 from torch.utils.data import DataLoader, DistributedSampler
 
 import wandb
@@ -91,6 +92,8 @@ def train(
     logger.info(
         f"Training with {len(train_dataloader)} batches of size {config.batch_size}."
     )
+    min_eval_loss = float("inf")
+    since_best = 0
     for epoch in range(start_epoch, config.max_epochs):
         if distributed_parameters["distributed"]:
             assert isinstance(train_dataloader.sampler, DistributedSampler)
@@ -166,6 +169,7 @@ def train(
             torch.distributed.all_reduce(stats, op=torch.distributed.ReduceOp.SUM)
             train_loss_sum, train_n = stats.tolist()
 
+        flag_tensor = torch.zeros(1).to(device)
         if distributed_parameters["rank"] == 0:
             model.eval()
             logger.info("Evaluating...")
@@ -191,7 +195,7 @@ def train(
                 train_loss = train_loss_sum / train_n
                 eval_loss = eval_loss_sum / eval_n
             # Log results
-            print(f"Epoch {epoch}\tLoss: {train_loss}\tEval loss: {eval_loss}")
+            logger.info(f"Epoch {epoch}\tLoss: {train_loss}\tEval loss: {eval_loss}")
             wandb.log(
                 {
                     "train/loss": train_loss,
@@ -211,15 +215,43 @@ def train(
                 },
                 models_folder / f"{run_id}.checkpoint.pt",
             )
+            if eval_loss < min_eval_loss:
+                min_eval_loss = eval_loss
+                since_best = 0
+                best_checkpoint_dir = models_folder / f"{run_id}.model"
+                (
+                    model.module if distributed_parameters["distributed"] else model
+                ).save_pretrained(best_checkpoint_dir)
+                tokenizer.save_pretrained(best_checkpoint_dir)
+                logger.info(f"Saved model to {best_checkpoint_dir.resolve()}")
+            else:
+                since_best += 1
+                if (
+                    (epoch + 1) >= config.min_epochs
+                    and config.early_stopping > 0
+                    and since_best >= config.early_stopping
+                ):
+                    logger.info(
+                        f"Early stopping. No improvements in the last {since_best} epochs"
+                    )
+                    flag_tensor += 1
+        # If we early stopped, broadcast break to all ranks
+        if distributed_parameters["distributed"]:
+            torch.distributed.all_reduce(flag_tensor, op=ReduceOp.SUM)
+        if flag_tensor.item() == 1:
+            break
 
     # Save final model and remove checkpoint
     if distributed_parameters["rank"] == 0:
         (models_folder / f"{run_id}.checkpoint.pt").unlink(missing_ok=True)
         final_checkpoint_dir = models_folder / f"{run_id}.model"
-        (
-            model.module if distributed_parameters["distributed"] else model
-        ).save_pretrained(final_checkpoint_dir)
-        tokenizer.save_pretrained(final_checkpoint_dir)
+        if not final_checkpoint_dir.exists():
+            # If the checkpoint already exists, we must have early stopped
+            # If not, make it now
+            (
+                model.module if distributed_parameters["distributed"] else model
+            ).save_pretrained(final_checkpoint_dir)
+            tokenizer.save_pretrained(final_checkpoint_dir)
         logger.info(f"Saved model to {final_checkpoint_dir.resolve()}")
 
         if config.new_hub_identifier:
